Move processors out of `holocron.ext` package
Because `holocron.ext` is a legacy we inherited from previous Holocron
implementation.
