+++ setup.py
@@ -55,21 +55,20 @@ setup(
         'console_scripts': [
             'holocron = holocron.main:main',
         ],
-        'holocron.ext.processors': [
-            'source = holocron.ext.processors.source:process',
-            'metadata = holocron.ext.processors.metadata:process',
-            'pipeline = holocron.ext.processors.pipeline:process',
-            'frontmatter = holocron.ext.processors.frontmatter:process',
-            'prettyuri = holocron.ext.processors.prettyuri:process',
-            'markdown = holocron.ext.processors.markdown:process',
-            'restructuredtext = '
-            '   holocron.ext.processors.restructuredtext:process',
-            'feed = holocron.ext.processors.feed:process',
-            'sitemap = holocron.ext.processors.sitemap:process',
-            'archive = holocron.ext.processors.archive:process',
-            'commit = holocron.ext.processors.commit:process',
-            'jinja2 = holocron.ext.processors.jinja2:process',
-            'when = holocron.ext.processors.when:process',
+        'holocron.processors': [
+            'source = holocron.processors.source:process',
+            'metadata = holocron.processors.metadata:process',
+            'pipeline = holocron.processors.pipeline:process',
+            'frontmatter = holocron.processors.frontmatter:process',
+            'prettyuri = holocron.processors.prettyuri:process',
+            'markdown = holocron.processors.markdown:process',
+            'restructuredtext = holocron.processors.restructuredtext:process',
+            'feed = holocron.processors.feed:process',
+            'sitemap = holocron.processors.sitemap:process',
+            'archive = holocron.processors.archive:process',
+            'commit = holocron.processors.commit:process',
+            'jinja2 = holocron.processors.jinja2:process',
+            'when = holocron.processors.when:process',
         ],
         'holocron.ext.commands': [
             'init = holocron.ext.commands.init:Init',
